gpio: fix gpio access functions
The dir_output function does not take the value of the
output (high/low) into account. This leads to situations where the
initial state is wrong.

Also, the gpio_set function passes a pointer to an 'int' and then uses
it as if it were a pointer to a 'u8' inside of pmc_write. This should
be avoided.